Inclusao do JPY, melhoria na transaction return ao pesquisar

+++ src/main/kotlin/br.com.jaya.coinconverter/model/CurrencySearchRequestDTO.kt
@@ -1,7 +1,9 @@
 package br.com.jaya.coinconverter.model
 
+import java.math.BigDecimal
+
 data class CurrencySearchRequestDTO(
    val currencyOrigin: String,
    val currencyDestiny: String,
-   val amount: Double
+   val amount: BigDecimal
 ){}
